Minor changes in nomenclature
