+++ app/controllers/assessment360_controller.rb
@@ -2,7 +2,7 @@ class Assessment360Controller < ApplicationController
   include GradesHelper
   include AuthorizationHelper
   include Scoring
-  include PenaltyHelper
+  include ScoreCalculationHelper
   # Added the @instructor to display the instructor name in the home page of the 360 degree assessment
   def action_allowed?
     current_user_has_ta_privileges?
